pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tts-stt-benchmark"
version = "0.1.0"
description = "Automated TTS/STT benchmark harness for latency and quality comparison"
requires-python = ">=3.11"
dependencies = [
    # HTTP / async
    "httpx>=0.27",
    "aiohttp>=3.9",
    "aiofiles>=23.2",
    # Provider SDKs
    "openai>=1.30",
    "deepgram-sdk>=3.6",
    "google-cloud-texttospeech>=2.16",
    "speechmatics-python>=1.11",
    # Audio processing
    "soundfile>=0.12",
    "numpy>=1.26",
    "scipy>=1.12",
    "pydub>=0.25",
    # Quality / NLP
    "jiwer>=3.0",
    # CLI & utils
    "click>=8.1",
    "rich>=13",
    "python-dotenv>=1.0",
    "pandas>=2.2",
    "tabulate>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5",
    "pytest-mock>=3.14",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
run_tts  = "tts_stt_benchmark.cli.run_tts:main"
run_stt  = "tts_stt_benchmark.cli.run_stt:main"
run_benchmark = "tts_stt_benchmark.cli.run_benchmark:main"
generate_report = "tts_stt_benchmark.cli.generate_report:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
